test: extract the common functions in test cases
